!!Clusterfying jobstore first attempt

+++ commons/types/src/main/java/dk/dbc/dataio/commons/types/Constants.java
@@ -20,7 +20,7 @@ public class Constants {
 
     // I'm introducing the invariant that submitter number 1 indicates missing value.
     public static final long MISSING_SUBMITTER_VALUE = 1;
-    public static final long PERSISTENCE_ID_LOWER_BOUND = 1L;
+    public static final int PERSISTENCE_ID_LOWER_BOUND = 1;
     public static final long PERSISTENCE_VERSION_LOWER_BOUND = 1L;
     public static final long CHUNK_ITEM_ID_LOWER_BOUND = -1L;
 }
